fix(workspace-mcp): journey projection, elicitation method, test layout
Fix P2: regenerate web/src/content/journeys/core.md — adds autonomous-dispatch
section added to packs/core/JOURNEY.md. Per web/AGENTS.md:49-53, tracked
generated files must be committed with the source change. Also synced
product-engineering journey (other changes in the current branch).

Fix P2: remove 'session/create_elicitation' from design.md:323 delivery-path
table. The ACP protocol uses a server-to-client MCP 'elicitation/create'
JSON-RPC request; the harness replies to the same request ID. No separate
'complete' call exists. The ACP SDK wrapper (unstable_createElicitation) is
noted as implementation detail, not the wire contract.

Fix P3: move TestFsmModeGuard (disk + git subprocess) to
tests/integration/test_workspace_mcp_git_fsm.py per packages/AGENTS.md:50-51
test layout rules. tests/test_workspace_mcp_git.py now holds only skip stubs.

Engine-Change-RFC: RFC-0078
